Change VCL::INT from C::int to C::long to gain more range on 64bit architectures.
NB:  VMOD writers need to review/revise use of the INT type.

Fixes #1204

+++ lib/libvcl/generate.py
@@ -482,7 +482,7 @@ vcltypes = {
 	'TIME':		"double",
 	'DURATION':	"double",
 	'BYTES':	"double",
-	'INT':		"int",
+	'INT':		"long",
 	'HEADER':	"const char *",
 }
